update axios calls
