+++ server/routes/users.js
@@ -1,7 +1,11 @@
 const express = require("express");
 const router = express.Router();
 
-const { validateUpdateUser, validateFileType } = require("../utils/validation");
+const {
+  validateUpdateUser,
+  validateFileType,
+  validatePassword,
+} = require("../utils/validation");
 const {
   getAllUsers,
   getPublicUser,
@@ -14,16 +18,18 @@ const {
   updateUserPhoto,
   updateUserAvatar,
   removeUserPhoto,
+  updatePassword,
 } = require("../controllers/usersController");
+const { requireAuth } = require("../utils/auth");
 
 router.get("/", getAllUsers);
-// router.post("/", validateSignup, createUser);
-router.get("/user", getSessionUser);
-router.patch("/user", validateUpdateUser, updateSessionUser);
-router.delete("/user", deleteSessionUser);
-router.patch("/user/photo", validateFileType, updateUserPhoto);
-router.delete("/user/photo", removeUserPhoto);
-router.patch("/user/avatar", updateUserAvatar);
+router.get("/user", requireAuth, getSessionUser);
+router.patch("/user", requireAuth, validateUpdateUser, updateSessionUser);
+router.delete("/user", requireAuth, deleteSessionUser);
+router.patch("/user/photo", requireAuth, validateFileType, updateUserPhoto);
+router.delete("/user/photo", requireAuth, removeUserPhoto);
+router.patch("/user/avatar", requireAuth, updateUserAvatar);
+router.patch("/user/password", requireAuth, validatePassword, updatePassword);
 router.patch("/forgot-password", updateUserResetPassword);
 router.patch("/reset-password/:resetPassword", resetUserPassword);
 router.post("/send-feedback", sendUserFeedback);
